Fixed a circular dependency caused by generateSuggestionsForFIlter being in the search-config file.
